Add tests based on typechecks
